handle conflicts and repetition
In order to prevent overwriting of existing values as a DOS attack
we disallow overwriting values entirely.

The keys in claimchain are all sha256 hashes
so there is no need to worry about conflicts.

In case one attempts to write the same data for a given key again
we return a 202 response code because the client can just resume.

If the data is different we return a 409
because the client needs to act upon this.